assignment2/app/requirements.txt
uuid
mysql-connector
rq
redis
flask